Break font handling into font controller
